requirements.txt
# XGuard Competition - 逆袭吧创二代
# 核心依赖

# 深度学习框架
torch>=2.1.0

# Transformers 生态
transformers>=4.45.0
accelerate>=0.27.0
peft>=0.10.0
datasets>=2.18.0

# 数据处理
scikit-learn>=1.3.0
numpy>=1.24.0

# 配置管理
pyyaml>=6.0

# 数据集加载 (可选, 二选一)
modelscope>=1.12.0

# 训练日志 (可选)
tensorboard>=2.14.0
wandb>=0.16.0
